Rename abstract.inf to abstract.infer
